Refactor: Plant 상세 ViewData #184
