+++ lib/features/plant/presentation/providers/plant_detail_view_provider.dart
@@ -1,16 +1,17 @@
 import 'package:commonplant_frontend/core/config/app_environment.dart';
-import 'package:commonplant_frontend/features/plant/domain/entities/plant_detail.dart';
 import 'package:commonplant_frontend/features/plant/presentation/fixtures/plant_detail_fixture.dart';
+import 'package:commonplant_frontend/features/plant/presentation/mappers/plant_detail_view_mapper.dart';
+import 'package:commonplant_frontend/features/plant/presentation/models/plant_detail_view_data.dart';
 import 'package:commonplant_frontend/features/plant/presentation/providers/plant_detail_remote_provider.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 typedef PlantDetailViewRequest = ({String plantId, String? placeCode});
 
 final plantDetailViewProvider =
-    Provider.family<
-      AsyncValue<PlantDetailFixtureData?>,
-      PlantDetailViewRequest
-    >((ref, request) {
+    Provider.family<AsyncValue<PlantDetailViewData?>, PlantDetailViewRequest>((
+      ref,
+      request,
+    ) {
       if (!ref.watch(useRemoteApiProvider)) {
         return AsyncData(ref.watch(plantLocalDetailViewProvider(request)));
       }
@@ -19,7 +20,7 @@ final plantDetailViewProvider =
     });
 
 final plantLocalDetailViewProvider =
-    Provider.family<PlantDetailFixtureData, PlantDetailViewRequest>((
+    Provider.family<PlantDetailViewData, PlantDetailViewRequest>((
       ref,
       request,
     ) {
@@ -27,28 +28,17 @@ final plantLocalDetailViewProvider =
     });
 
 final plantRemoteDetailViewProvider =
-    FutureProvider.family<PlantDetailFixtureData?, PlantDetailViewRequest>((
+    FutureProvider.family<PlantDetailViewData?, PlantDetailViewRequest>((
       ref,
       request,
     ) async {
-      final fixture = ref.watch(plantLocalDetailViewProvider(request));
+      final fallback = ref.watch(plantLocalDetailViewProvider(request));
       final detail = await ref.watch(
         remotePlantDetailProvider(request.plantId).future,
       );
 
-      return applyRemotePlantDetailToFixture(fixture: fixture, detail: detail);
+      return mapPlantDetailToViewData(fallback: fallback, detail: detail);
     }, retry: (retryCount, error) => null);
-
-PlantDetailFixtureData? applyRemotePlantDetailToFixture({
-  required PlantDetailFixtureData fixture,
-  required PlantDetail detail,
-}) {
-  if (detail.name.trim().isEmpty) {
-    return null;
-  }
-
-  return fixture.applyRemote(detail);
-}
 
 void invalidatePlantDetailView(WidgetRef ref, PlantDetailViewRequest request) {
   ref.invalidate(remotePlantDetailProvider(request.plantId));
