+++ test/features/plant/presentation/fixtures/plant_detail_fixture_test.dart
@@ -1,4 +1,3 @@
-import 'package:commonplant_frontend/features/plant/domain/entities/plant_detail.dart';
 import 'package:commonplant_frontend/features/plant/presentation/fixtures/plant_detail_fixture.dart';
 import 'package:flutter_test/flutter_test.dart';
 
@@ -15,26 +14,6 @@ void main() {
       expect(detail.wateringCycleLabel, '10 Day');
       expect(detail.memos, hasLength(4));
       expect(detail.memos.first.author, '커먼플랜트');
-    });
-
-    test('remote detail을 적용할 때 없는 값은 fallback을 유지한다', () {
-      final detail = plantDetailFixture(placeCode: 'fallback-place')
-          .applyRemote(
-            const PlantDetail(
-              id: 'plant-remote',
-              name: '필로덴드론',
-              placeName: '거실 정원',
-              lastWateredDate: '2026.05.25',
-            ),
-          );
-
-      expect(detail.placeCode, 'fallback-place');
-      expect(detail.placeName, '거실 정원');
-      expect(detail.name, '필로덴드론');
-      expect(detail.species, 'Monstera deliciosa');
-      expect(detail.lastWateredDate, '2026.05.25');
-      expect(detail.dDayLabel, 'D-3');
-      expect(detail.memos, hasLength(4));
     });
   });
 }
